rename generator package to model
