+++ cmd/main.go
@@ -7,7 +7,7 @@ import (
 	"os"
 
 	tea "github.com/charmbracelet/bubbletea"
-	"github.com/liioan/faek/internal/generator"
+	"github.com/liioan/faek/internal/model"
 	"github.com/liioan/faek/internal/utils"
 )
 
@@ -18,19 +18,25 @@ func main() {
 	flag.BoolVar(&config, "c", false, "enter configuration mode")
 	flag.Parse()
 
+	steps := []model.Step{
+
+	}
+
 	if config {
-		fmt.Println("config")
-		return
+		steps = []model.Step{
+			*model.NewListStep("Choose your default output style", false, {});
+		}
+	} else {
+	steps = []model.Step{
+		*model.NewTextStep("What will the array be called? (default: arr)", "e.g. users", false),
+		*model.NewTextStep("Write your field (to continue press enter without input)", "e.g. name string", true),
+		*model.NewTextStep("Create type for your object? (default: no type, input: type name)", "e.g. Post", false),
+		*model.NewTextStep("How many items will be in this array (default 5)", "e.g. 5", false),
+	}
 	}
 
-	steps := []generator.Step{
-		*generator.NewTextStep("What will the array be called? (default: arr)", "e.g. users", false),
-		*generator.NewTextStep("Write your field (to continue press enter without input)", "e.g. name string", true),
-		*generator.NewTextStep("Create type for your object? (default: no type, input: type name)", "e.g. Post", false),
-		*generator.NewTextStep("How many items will be in this array (default 5)", "e.g. 5", false),
-	}
 
-	model := generator.NewModel(steps, false)
+	model := model.NewModel(steps, false)
 
 	file, err := tea.LogToFile("debug.log", "debug")
 	if err != nil {
