+++ internal/model/model.go
@@ -1,4 +1,4 @@
-package generator
+package model
 
 import (
 	"strings"
